Tracing: Move tracing thread listeners to trace-common
Previously, only tracemode projections used these listeners, meaning
other tracemodes did not get EP information for resumptions of
[threaded] EPs. Now all tracemodes will automatically get the correct
data.
