+++ src/ck-perf/trace-projections.C
@@ -110,63 +110,6 @@ void _createTraceprojections(char **argv)
   CkpvAccess(_traces)->addTrace(CkpvAccess(_trace));
   if (CkMyPe()==0) CkPrintf("Charm++: Tracemode Projections enabled.\n");
 }
- 
-/* ****** CW TEMPORARY LOCATION ***** Support for thread listeners */
-
-struct TraceThreadListener {
-  struct CthThreadListener base;
-  int event;
-  int msgType;
-  int ep;
-  int srcPe;
-  int ml;
-  CmiObjId idx;
-};
-
-static void traceThreadListener_suspend(struct CthThreadListener *l)
-{
-  TraceThreadListener *a=(TraceThreadListener *)l;
-  /* here, we activate the appropriate trace codes for the appropriate
-     registered modules */
-  traceSuspend();
-}
-
-static void traceThreadListener_resume(struct CthThreadListener *l)
-{
-  TraceThreadListener *a=(TraceThreadListener *)l;
-  /* here, we activate the appropriate trace codes for the appropriate
-     registered modules */
-  _TRACE_BEGIN_EXECUTE_DETAILED(a->event,a->msgType,a->ep,a->srcPe,a->ml,
-				CthGetThreadID(a->base.thread), NULL);
-  a->event=-1;
-  a->srcPe=CkMyPe(); /* potential lie to migrated threads */
-  a->ml=0;
-}
-
-static void traceThreadListener_free(struct CthThreadListener *l)
-{
-  TraceThreadListener *a=(TraceThreadListener *)l;
-  delete a;
-}
-
-void TraceProjections::traceAddThreadListeners(CthThread tid, envelope *e)
-{
-#if CMK_TRACE_ENABLED
-  /* strip essential information from the envelope */
-  TraceThreadListener *a= new TraceThreadListener;
-  
-  a->base.suspend=traceThreadListener_suspend;
-  a->base.resume=traceThreadListener_resume;
-  a->base.free=traceThreadListener_free;
-  a->event=e->getEvent();
-  a->msgType=e->getMsgtype();
-  a->ep=e->getEpIdx();
-  a->srcPe=e->getSrcPe();
-  a->ml=e->getTotalsize();
-
-  CthAddListener(tid, (CthThreadListener *)a);
-#endif
-}
 
 void LogPool::openLog(const char *mode)
 {
